UX: Display confirmation message after subscribing (#2)
Currently, when a user subscribes to the newsletter via the banner at top of the site, the banner disappears as soon as the subscribe button is clicked. This behavior isn't ideal from a UX perspective because there's no confirmation that the user action has done anything.

This commit improves the subscribe UX by displaying a confirmation message after the user clicks the subscribe button. It also moves the 'Manage preferences' link from the initial banner state to the confirmation message.

Internal topic: t/96002/31.